requirements.txt
pandas
matplotlib
python-dotenv
google-generativeai